Correctly use Spring 'Inversion of Control' to inject service classes

+++ src/main/java/net/andreweast/services/ga/service/GeneticAlgorithmSerializer.java
@@ -7,10 +7,12 @@ import net.andreweast.services.data.model.Job;
 import net.andreweast.services.data.model.Schedule;
 import net.andreweast.services.ga.dataaccess.JobDao;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 /**
  * Puts all information from a completed Genetic Algorithm run back into the database
  */
+@Service
 public class GeneticAlgorithmSerializer {
     @Autowired
     private ScheduleRepository scheduleRepository;
@@ -29,11 +31,17 @@ public class GeneticAlgorithmSerializer {
         // Finally, mark this Job as being done by deleting the Job record, setting the Schedule's job foreign key to null
         Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(DataNotFoundException::new);
 
-        Job job = jobRepository.findById(schedule.getJob().getJobId()).orElseThrow(DataNotFoundException::new);
+        // Only attempt to delete job if this Schedule record has a job
+        // Since this is wired to an idempotent REST service, if it doesn't exist (already deleted or never did), then no need for an error!
+        if (schedule.getJob() == null) {
+            System.out.println("Schedule record has no job to delete, schedule_id=" + scheduleId); // TODO: Logger
+        } else {
+            Job job = jobRepository.findById(schedule.getJob().getJobId()).orElseThrow(DataNotFoundException::new);
 
-        schedule.setJob(null);
-        scheduleRepository.save(schedule);
+            schedule.setJob(null);
+            scheduleRepository.save(schedule);
 
-        jobRepository.delete(job);
+            jobRepository.delete(job);
+        }
     }
 }
